Déplacement de TripProposal dans un nouveau composant
